+++ src/pages/projectSettings/HeaderButtons.tsx
@@ -66,9 +66,11 @@ export const HeaderButtons: React.VFC<Props> = ({ id, projectType, tab }) => {
       dispatchToast.success("Successfully updated project");
 
       if (identifier !== newIdentifier) {
-        navigate(getProjectSettingsRoute(newIdentifier, tab), {
-          replace: true,
-        });
+        setTimeout(() => {
+          navigate(getProjectSettingsRoute(newIdentifier, tab), {
+            replace: true,
+          });
+        }, 500);
       }
     },
     onError(err) {
